# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ardana Smart Plant Care and Farm Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Plant {
  id: number;
  name: string;
  /** @nullable */
  species?: string | null;
  /** crop | tree | plant | flower | herb | shrub */
  type: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  plantedDate?: string | null;
  /** @nullable */
  lastWateredDate?: string | null;
  /** @nullable */
  lastFertilizedDate?: string | null;
  /** @nullable */
  lastPrunedDate?: string | null;
  /** healthy | moderate | poor | unknown */
  healthStatus: string;
  /** @nullable */
  wateringIntervalDays?: number | null;
  /** @nullable */
  fertilizingIntervalDays?: number | null;
  /** @nullable */
  pruningIntervalDays?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Reminder {
  id: number;
  plantId: number;
  plantName: string;
  /** watering | fertilizing | pruning | spraying | harvesting | other */
  type: string;
  scheduledDate: string;
  /**
     * HH:MM time string, e.g. "09:00"
     * @nullable
     */
  scheduledTime: string | null;
  completed: boolean;
  isCustom: boolean;
  /**
     * Interval in days for custom recurring reminders; null means does not repeat
     * @nullable
     */
  recurrenceDays: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PlantDetail {
  id: number;
  name: string;
  /** @nullable */
  species?: string | null;
  type: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  plantedDate?: string | null;
  /** @nullable */
  lastWateredDate?: string | null;
  /** @nullable */
  lastFertilizedDate?: string | null;
  /** @nullable */
  lastPrunedDate?: string | null;
  healthStatus: string;
  /** @nullable */
  wateringIntervalDays?: number | null;
  /** @nullable */
  fertilizingIntervalDays?: number | null;
  /** @nullable */
  pruningIntervalDays?: number | null;
  createdAt: string;
  updatedAt: string;
  reminders: Reminder[];
}

export interface PlantInput {
  /** @minLength 1 */
  name: string;
  species?: string;
  type: string;
  location?: string;
  notes?: string;
  photoUrl?: string;
  plantedDate?: string;
  healthStatus?: string;
  wateringIntervalDays?: number;
  fertilizingIntervalDays?: number;
  pruningIntervalDays?: number;
}

export interface PlantUpdate {
  /** @minLength 1 */
  name?: string;
  species?: string;
  type?: string;
  location?: string;
  notes?: string;
  photoUrl?: string;
  plantedDate?: string;
  lastWateredDate?: string;
  lastFertilizedDate?: string;
  lastPrunedDate?: string;
  healthStatus?: string;
  wateringIntervalDays?: number;
  fertilizingIntervalDays?: number;
  pruningIntervalDays?: number;
}

export interface ReminderInput {
  plantId: number;
  type: string;
  scheduledDate: string;
  /** HH:MM time string, e.g. "09:00" */
  scheduledTime?: string;
  /** Interval in days for recurring reminders; omit for one-time */
  recurrenceDays?: number;
  notes?: string;
}

export interface ReminderUpdate {
  type?: string;
  scheduledDate?: string;
  /** HH:MM time string, e.g. "09:00" */
  scheduledTime?: string;
  /** Interval in days for custom recurring reminders. For AI-generated care reminders (watering/fertilizing/pruning), this updates the plant's care interval instead. */
  recurrenceDays?: number;
  completed?: boolean;
  notes?: string;
}

export type DashboardStatsByType = {[key: string]: number};

export type DashboardStatsByHealth = {[key: string]: number};

export interface DashboardStats {
  totalPlants: number;
  byType: DashboardStatsByType;
  byHealth: DashboardStatsByHealth;
  upcomingRemindersCount: number;
  overdueRemindersCount: number;
  /** Number of plants watered today or yesterday */
  recentlyWatered: number;
  /** Number of plants watered today */
  wateredToday: number;
  /** Number of plants watered yesterday */
  wateredYesterday: number;
  needsAttention: number;
}

export interface PlantImageInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type, e.g. image/jpeg */
  mimeType: string;
  /** User interface language for the AI response (e.g. en, ar, fr, es, pt) */
  language?: string;
}

export interface PlantIdentificationResult {
  /** @nullable */
  species: string | null;
  /** @nullable */
  commonName: string | null;
  /** @nullable */
  family: string | null;
  /** @nullable */
  confidence: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  wateringRequirements: string | null;
  /** @nullable */
  growingConditions: string | null;
  /** @nullable */
  fertilizers: string | null;
  /** @nullable */
  careRecommendations: string | null;
  /** @nullable */
  sunlight: string | null;
  /** @nullable */
  soilType: string | null;
  /** @nullable */
  error: string | null;
}

export interface DiseaseDetectionResult {
  isHealthy: boolean;
  /** @nullable */
  diseaseName: string | null;
  /** @nullable */
  confidence: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  causes: string | null;
  /** @nullable */
  symptoms: string | null;
  /** @nullable */
  treatments: string | null;
  /** @nullable */
  products: string | null;
  /** @nullable */
  preventiveMeasures: string | null;
  /** @nullable */
  urgency: string | null;
  /** @nullable */
  error: string | null;
}

export interface WeatherCurrent {
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  windDir: string;
  windDegree: number;
  windGust: number;
  weatherCode: number;
  weatherDescription: string;
  weatherIcon: string;
  isDay: boolean;
  precipitation: number;
  pressure: number;
  visibility: number;
  cloudCover: number;
  uvIndex: number;
  airQualityIndex: number;
}

export interface WeatherHourly {
  time: string;
  temperature: number;
  feelsLike: number;
  precipitation: number;
  chanceOfRain: number;
  weatherCode: number;
  weatherDescription: string;
  humidity: number;
  windSpeed: number;
  windDir: string;
  uvIndex: number;
  cloudCover: number;
}

export interface WeatherDaily {
  date: string;
  maxTemp: number;
  minTemp: number;
  avgHumidity: number;
  precipitation: number;
  chanceOfRain: number;
  maxWindSpeed: number;
  weatherCode: number;
  weatherDescription: string;
  weatherIcon: string;
  sunrise: string;
  sunset: string;
  moonPhase: string;
  uvIndex: number;
  hourly: WeatherHourly[];
}

export interface WeatherData {
  locationName: string;
  lat: number;
  lon: number;
  timezone?: string;
  current: WeatherCurrent;
  hourly: WeatherHourly[];
  daily: WeatherDaily[];
  /** AI-generated care recommendation based on weather */
  recommendation: string;
}

export type ListPlantsParams = {
type?: string;
healthStatus?: string;
search?: string;
};

export type ListRemindersParams = {
completed?: string;
type?: string;
/**
 * If "true", return only upcoming/overdue reminders
 */
upcoming?: string;
};

export type GetWeatherParams = {
lat: number;
lon: number;
locationName?: string;
/**
 * User interface language for localized weather descriptions
 */
language?: string;
};