# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ardana Smart Plant Care and Farm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/identify-plant:
    post:
      operationId: identifyPlant
      tags: [ai]
      summary: Identify a plant from an uploaded image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantImageInput"
      responses:
        "200":
          description: Plant identification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantIdentificationResult"
  /ai/detect-disease:
    post:
      operationId: detectDisease
      tags: [ai]
      summary: Detect plant disease from an uploaded image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantImageInput"
      responses:
        "200":
          description: Disease detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseDetectionResult"
...
components:
  schemas:
...
    PlantImageInput:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data
        mimeType:
          type: string
          description: Image MIME type, e.g. image/jpeg
        language:
          type: string
          description: User interface language for the AI response (e.g. en, ar, fr, es, pt)
          default: en
      required:
        - imageBase64
        - mimeType
    PlantIdentificationResult:
      type: object
      properties:
        species:
          type: ["string", "null"]
        commonName:
          type: ["string", "null"]
        family:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        wateringRequirements:
          type: ["string", "null"]
        growingConditions:
          type: ["string", "null"]
        fertilizers:
          type: ["string", "null"]
        careRecommendations:
          type: ["string", "null"]
        sunlight:
          type: ["string", "null"]
        soilType:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
      required:
        - species
        - commonName
        - family
        - confidence
        - description
        - wateringRequirements
        - growingConditions
        - fertilizers
        - careRecommendations
        - sunlight
        - soilType
        - error
    DiseaseDetectionResult:
      type: object
      properties:
        isHealthy:
          type: boolean
        diseaseName:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        causes:
          type: ["string", "null"]
        symptoms:
          type: ["string", "null"]
        treatments:
          type: ["string", "null"]
        products:
          type: ["string", "null"]
        preventiveMeasures:
          type: ["string", "null"]
        urgency:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
      required:
        - isHealthy
        - diseaseName
        - confidence
        - description
        - causes
        - symptoms
        - treatments
        - products
        - preventiveMeasures
        - urgency
        - error
...