# lib/api-zod/src/generated/types/plantImageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Ardana Smart Plant Care and Farm Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PlantImageInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type, e.g. image/jpeg */
  mimeType: string;
  /** User interface language for the AI response (e.g. en, ar, fr, es, pt) */
  language?: string;
}